refactor(workflow): split PR readiness evidence tests
